Rename profile tests file, apply code suggestions
